IPC, SCICLIENT: Add support for AM62X
SCICLIENT driver ported to AM62x.
IPC driver update to support AM62X SoC.
This version of IPC will be used to build for QNX.

+++ packages/ti/drv/mailbox/src/mailbox_internal.h
@@ -265,13 +265,13 @@ typedef struct Mailbox_Driver_t
      */
     Mailbox_RemoteCfg     *remoteCfgPtr;
     /**
-     * @brief   Base Address of the Tx Mailbox for this instance (used for am65xx/am64x/j721e/j7200)
+     * @brief   Base Address of the Tx Mailbox for this instance (used for am65xx/am64x/j721e/j7200/am62x)
      */
-    uint32_t              baseAddrTx;
+    uintptr_t             baseAddrTx;
     /**
-     * @brief   Base Address of the Rx Mailbox for this instance (used for am65xx/am64x/j721e/j7200)
+     * @brief   Base Address of the Rx Mailbox for this instance (used for am65xx/am64x/j721e/j7200/am62x)
      */
-    uint32_t              baseAddrRx;
+    uintptr_t             baseAddrRx;
 }Mailbox_Driver;
 
 /*!
